refactor(ui5-link): migrate to TS
