+++ packages/main/src/Link.hbs
@@ -4,7 +4,7 @@
 	href="{{parsedRef}}"
 	target="{{target}}"
 	rel="{{_rel}}"
-	tabindex="{{tabIndex}}"
+	tabindex="{{effectiveTabIndex}}"
 	?disabled="{{disabled}}"
 	aria-label="{{ariaLabelText}}"
 	aria-haspopup="{{accessibilityAttributes.hasPopup}}"
